binance_bot/requirements.txt
python-binance 
pandas_ta
websocket_client